Dual System
- yieldbuddy folder
- outdoor system 4 water tanks, 5 pods, T/Hum, Light, Soil Hum, 2 pH
- 2 TDS, CO2, 4 water pump
- 17 sensor values stored in database

- yieldbuddy2 folder
- Indoor Mini Greenhouse, 3 shelf, T/Hum, Light, Soil Hum, 2 pH
- 2 TDS, CO2, 4 water pumps
- 17 sensor readings stored in database
Both systems are running on separate USB to TTL converters connected to
Ubuntu 14.04 SQliteAMP server.

- all sensors are working
- all setpoints are working
- RH control is working via fan on Relay6
- Relay1-4 connected to tank1-4 pumps
- 17 sensor readings stored in database

Current things that need fixing

- the startup deamon and scripts need reworking. On system boot the
script will start the yieldbuddy.py in yieldbuddy folder and keep it
running. The yieldbuddy2.py script needs to be started manual to get
both systems logging data.

The current test script just looks for the "yieldbuddy" string when
considering if the script is running or not. This doesn't work in a two
script system since both scripts contain the string.

First idea for fix is to change the names to something distinct so the
deamon can tell if one or the other is running and act accordingly.

I am thinking it would be better to track this with /var/run entries and
get specific pid control on if the separate yieldbuddy.py yieldbuddy2.py
scripts are running. I think this would be a good place to also use
fuser to make sure that all old locks are removed before starting. Going
to do some research on what the best practice is for keeping a py script
up and running.